added email domain check

+++ SMTP.py
@@ -9,10 +9,10 @@ import thread
 
 
 #start a new file for this session and change the state of the HELO 
-def HELO(args, socket, client_address, state):
+def HELO(args, s, client_address, state):
     fileName = str(client_address[1]) + '.txt'
     if len(args) != 2:
-      socket.send("501 Syntax: HELO hostname \n")
+      s.send("501 Syntax: HELO hostname \n")
       return
     #check if helo has been sent before
     if state['HELO'] == False:
@@ -21,7 +21,7 @@ def HELO(args, socket, client_address, state):
       state['HELO'] = True
       state['file'] = client_address[1]
       state['domain'] = args[1]
-      socket.send("250 "+ str(client_address[1]) + "OK \n")
+      s.send("250 "+ str(client_address[1]) + "OK \n")
     #if helo sent before reset all state and delete old file
     else:
       open(fileName, 'w').close()
@@ -32,20 +32,20 @@ def HELO(args, socket, client_address, state):
       state['RCPT'] = False
 
       state['HELO'] = True
-      socket.send("250 "+ str(client_address[1]) + " OK \n")
+      s.send("250 "+ str(client_address[1]) + " OK \n")
 
 #start the mail transaction after checking the sessions has be inititalized
-def MAIL(args, socket, client_address, state):
+def MAIL(args, s, client_address, state):
     fileName = str(state['file']) + '.txt'
     #first check that helo has been sent
     if state['HELO'] == False:
-      socket.send("503 5.5.1 Error: send HELO/EHLO first \n")
+      s.send("503 5.5.1 Error: send HELO/EHLO first \n")
     else:
       #make sure it's not a nested mail command
       if state['MAIL'] == False:
         #check if the arguments are provided
         if len(args) != 2:
-          socket.send("501 5.5.4 Syntax: MAIL FROM:<address> \n")
+          s.send("501 5.5.4 Syntax: MAIL FROM:<address> \n")
           return
         checkSyntax = re.match("FROM:<\w+@\w+\.\w+>", args[1], re.IGNORECASE)
         if(checkSyntax):
@@ -54,7 +54,7 @@ def MAIL(args, socket, client_address, state):
             with open(fileName, 'a') as the_file:
               the_file.write(" ".join(args) + "\n")
             state['MAIL'] = True
-            socket.send("250 2.1.0 Ok \n")
+            s.send("250 2.1.0 Ok \n")
           else:
               state['file'] = state['file'] + 1
               fileName = str(state['file']) + '.txt'
@@ -62,19 +62,17 @@ def MAIL(args, socket, client_address, state):
                 the_file.write("helo " + state['domain'] + "\n")
                 the_file.write(" ".join(args) + "\n")
               state['MAIL'] = True
-              socket.send("250 2.1.0 Ok \n")
-
-
+              s.send("250 2.1.0 Ok \n")
         else:
-          socket.send("501 5.1.7 Bad sender address syntax \n")
+          s.send("501 5.1.7 Bad sender address syntax \n")
       else:
-        socket.send("503 5.5.1 Error: nested MAIL command \n")
+        s.send("503 5.5.1 Error: nested MAIL command \n")
     
-def RCPT(args, socket, client_address, state):
+def RCPT(args, s, client_address, state):
   #check if a mail transaction has begon and helo is initiatied
   if state['MAIL'] == True and state['HELO'] == True:
     if len(args) != 2:
-      socket.send("501 5.5.4 Syntax: RCPT TO:<address> \n")
+      s.send("501 5.5.4 Syntax: RCPT TO:<address> \n")
       return
     #check the format of the email is valid
     checkSyntax = re.match("TO:<\w+@\w+\.\w+>", args[1], re.IGNORECASE)
@@ -85,40 +83,40 @@ def RCPT(args, socket, client_address, state):
       with open(fileName, 'a') as the_file:
         the_file.write(" ".join(args) + "\n")
       state['RCPT'] = True
-      socket.send("250 2.1.5 Ok \n")
+      s.send("250 2.1.5 Ok \n")
     else:
-      socket.send("501 5.1.3 Bad recipient address syntax \n")
+      s.send("501 5.1.3 Bad recipient address syntax \n")
   else:
-    socket.send("503 5.5.1 Error: need MAIL command \n")
+    s.send("503 5.5.1 Error: need MAIL command \n")
 
-def DATA(args, socket, client_address, state):
+def DATA(args, s, client_address, state):
   fileName = str(state['file']) + '.txt'
   if state['MAIL'] == True and state['HELO'] == True and state['RCPT'] == True:
-    socket.send("354 End data with <CR><LF>.<CR><LF> \n")
-    data = recieveData(socket)
+    s.send("354 End data with <CR><LF>.<CR><LF> \n")
+    data = recieveData(s)
     with open(fileName, 'a') as the_file:
       the_file.write("data \n")
       the_file.write(data + "\n")
       the_file.write("quit \n")
     state['MAIL'] = False
     state['RCPT'] = False
-    socket.send("Qeued " + str(state['file']) + " \n")
+    s.send("Qeued " + str(state['file']) + " \n")
     state['data'] = True
     thread.start_new_thread(relayData,(state['file'], state))
   elif state['MAIL'] == True and state['RCPT'] == False:
-    socket.send("554 5.5.1 Error: no valid recipients \n")
+    s.send("554 5.5.1 Error: no valid recipients \n")
   else:
-    socket.send("503 5.5.1 Error: need RCPT command \n")
+    s.send("503 5.5.1 Error: need RCPT command \n")
 
-def QUIT(args, socket, client_address, state):
+def QUIT(args, s, client_address, state):
   state['loop'] = False
-  socket.send("221 2.0.0 Bye \n")
-  socket.close()
+  s.send("221 2.0.0 Bye \n")
+  s.close()
 
-def VRFY(args, socket, client_address, state):
-  socket.send("252  Cannot VRFY user \n")
+def VRFY(args, s, client_address, state):
+  s.send("252  Cannot VRFY user \n")
 
-def RSET(args, socket, client_address, state):
+def RSET(args, s, client_address, state):
   fileName = str(state['file'] + '.txt')
   with open(fileName) as f:
     first_line = f.readline()
@@ -127,47 +125,56 @@ def RSET(args, socket, client_address, state):
     the_file.write(first_line)
   state['MAIL'] = False
   state['RCPT'] = False
-  socket.send("250 OK \n")
+  s.send("250 OK \n")
 
-
-def relayData(client_address, state):
-  filename = str(client_address) + '.txt'
-  # The remote host
-  domain = re.search("@[\w.]+", state['recipient'])
+def findMXServer(email):
+  domain = re.search("@[\w.]+", email)
   domain = domain.group()
   domain = domain[1:]
-  mailExchangeServers = dns.resolver.query(domain, 'MX')
+  try:
+    mailExchangeServers = dns.resolver.query(domain, 'MX')
+  except:
+    print "no answer \n"
+    return
   lowestPref = ""
   pref = mailExchangeServers[0].preference
   for rdata in mailExchangeServers:
     if rdata.preference <= pref:
       lowestPref = rdata.exchange.__str__()
   lowestPref = lowestPref[:-1]
-  print lowestPref
-  HOST = lowestPref
-  PORT = 25              # email port
-  s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-  s.connect((HOST, PORT))
-  data = s.recv(1024)
-  print('Received', repr(data))
-  with open(filename) as fp:
-    for line in fp:
-      print('sent', repr(line))
-      s.sendall(line)
-      if line == ".\\n":
-        data = s.recv(1024)
-        print('Received', repr(data))
-  os.remove(filename)
-  data = s.recv(1024)
-  print('Received', repr(data))
+  return lowestPref
+
+def relayData(client_address, state):
+  filename = str(client_address) + '.txt'
+  # The remote host
+  HOST = findMXServer(state['recipient'])
+  print "the host is ", HOST
+  if HOST:
+    PORT = 25              # email port
+    s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    s.connect((HOST, PORT))
+    data = s.recv(1024)
+    print('Received', repr(data))
+    with open(filename) as fp:
+      for line in fp:
+        print('sent', repr(line))
+        s.sendall(line)
+        if line == ".\\n":
+          data = s.recv(1024)
+          print('Received', repr(data))
+    os.remove(filename)
+    data = s.recv(1024)
+    print('Received', repr(data))
+  else:
+    print "Host not found \n"
 
 #state['recipient'] = "FROM:<nsaffour@gmail.com>"
 #relayData(33345)
 
-def recieveData(socket):
+def recieveData(s):
     fragments = []
     while True: 
-      line = linesplit(socket)
+      line = linesplit(s)
       line = line + '\n'
       fragments.append(line)
       if line == ".\r\n":
@@ -182,32 +189,32 @@ dispatch = {
     'vrfy': VRFY,
     'rest': RSET
 }
-def process_network_command(command, args, socket, client_address, state):
+def process_network_command(command, args, s, client_address, state):
   command = command.lower()
   try:
-    dispatch[command](args, socket, client_address, state)
+    dispatch[command](args, s, client_address, state)
   except KeyError:
-    socket.send("502 5.5.2 Error: command not recognized \n")
+    s.send("502 5.5.2 Error: command not recognized \n")
 
-def linesplit(socket):
+def linesplit(s):
     #add timeout to the connection if no commands are recieved
-    socket.settimeout(300)
-    buffer = socket.recv(4096)
+    s.settimeout(300)
+    buffer = s.recv(4096)
     #remove timeout if commands are recieved
-    socket.settimeout(None)
+    s.settimeout(None)
     buffering = True
     while buffering:
         if "\n" in buffer:
             (line, buffer) = buffer.split("\n", 1)
             return line
         else:
-            more = socket.recv(4096)
+            more = s.recv(4096)
             if not more:
                 buffering = False
             else:
                 buffer += more
 
-def handleClient(socket, client_address):
+def handleClient(s, client_address):
   state = {
   'HELO': False,
   'MAIL': False,
@@ -219,7 +226,7 @@ def handleClient(socket, client_address):
   'domain': ""
   }
   try:
-    connection.send("220 SMTP Nour 1.0 \n")
+    s.send("220 SMTP Nour 1.0 \n")
     print >>sys.stderr, 'connection from', client_address
     # Receive the data in small chunks 
     while state['loop']:
@@ -229,7 +236,7 @@ def handleClient(socket, client_address):
         process_network_command(args[0], args, connection, client_address, state)
   finally:
       # Clean up the connection
-      connection.close()
+      s.close()
 
 # Create a TCP/IP socket
 sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
